start of 'recent changes' screen
I'm still unclear on precisely what the experience should be here. We
probably don't need all of a version control system. The goal is just to
be able to answer the question, "what did I change recently that caused
things to break?"

For now let's just start with letting people see past versions.