zet.tlv: refine the initial UX some more
Conundrum:

I want the app to provide some sample data for people to interact with.

However, I also want them to be able to delete it to make it their own.

BUT I'm not yet convinced the app should allow deletion in general.

Besides, deleting stuff robustly and reliably is a whole other level of
programming.

Solution:

Make people edit the app to clear the initial sample zettels.

This has the "beneficial" side-effect of getting them editing apps on
Teliva.